# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeClash - Real-time competitive coding battle platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/{username}/history:
    get:
      operationId: getUserBattleHistory
      tags: [users]
      summary: Get user battle history
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BattleHistoryEntry"
  /users/sync-codeforces:
    post:
      operationId: syncCodeforcesHandle
      tags: [users]
      summary: Sync user's Codeforces handle data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CfSyncRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Rooms ──
...
components:
  schemas:
...
    UserRegistration:
      type: object
      required: [username, password, cfHandle]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        cfHandle:
          type: string
    UserLogin:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── Users ──
    User:
      type: object
      required: [id, username, cfHandle, rating, rank, battleWins, battleLosses, totalBattles, winStreak, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        cfHandle:
          type: string
        rating:
          type: integer
        maxRating:
          type: integer
        rank:
          type: string
        maxRank:
          type: string
        contribution:
          type: integer
        battleWins:
          type: integer
        battleLosses:
          type: integer
        totalBattles:
          type: integer
        winStreak:
          type: integer
        favoriteTopics:
          type: array
          items:
            type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        favoriteTopics:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
    CfSyncRequest:
      type: object
      required: [cfHandle]
      properties:
        cfHandle:
          type: string
    BattleHistoryEntry:
      type: object
      required: [battleId, battleType, result, solveTime, attempts, problemTitle, problemRating, opponent, createdAt]
      properties:
        battleId:
          type: integer
        battleType:
          type: string
        result:
          type: string
          enum: [win, loss, draw]
        solveTime:
          type: ["integer", "null"]
        attempts:
          type: integer
        problemTitle:
          type: string
        problemRating:
          type: integer
        opponent:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Rooms ──
...