# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeClash - Real-time competitive coding battle platform
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  cfHandle: string;
  rating: number;
  maxRating?: number;
  rank: string;
  maxRank?: string;
  contribution?: number;
  battleWins: number;
  battleLosses: number;
  totalBattles: number;
  winStreak: number;
  favoriteTopics?: string[];
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}